Move temp mapping file handle to a separate header.
